Move drawing event to IconPane MVC
