+++ griffon-app/resources/org/laeq/editor/iconPane.fxml
@@ -3,6 +3,7 @@
 
 <?import javafx.scene.layout.AnchorPane?>
 <?import org.laeq.editor.IconPane?>
+<?import org.laeq.editor.DrawingPane?>
 <AnchorPane fx:id="iconAnchor" xmlns="http://javafx.com/javafx/11.0.10" xmlns:fx="http://javafx.com/fxml/1"
             fx:controller="org.laeq.editor.IconPaneController"
             style="-fx-border-color: green; -fx-border-width: 2"
@@ -11,6 +12,7 @@
             AnchorPane.rightAnchor="0.0"
             AnchorPane.topAnchor="0.0">
     <children>
+        <DrawingPane fx:id="drawingPane" layoutX="0.0" layoutY="0.0"  style="-fx-border-color: #FF0F0F; -fx-border-width: 5; " />
         <IconPane fx:id="container" style="-fx-border-color: orange; -fx-border-width: 2; " />
     </children>
 </AnchorPane>
