+++ griffon-app/views/org/laeq/editor/ImageControlsView.java
@@ -55,7 +55,7 @@ public class ImageControlsView extends AbstractJavaFXGriffonView {
         stage.getIcons().add( getImage("favicon-32x32.png"));
         stage.setScene(init());
         stage.sizeToScene();
-        stage.setAlwaysOnTop(false);
+        stage.setAlwaysOnTop(true);
         getApplication().getWindowManager().attach("image_controls", stage);
         getApplication().getWindowManager().show("image_controls");
